refactor: migrate inspectConfig functionality to a dedicated file
